requirements.txt
pandas
psycopg2-binary
SQLAlchemy
python-dotenv